ingest/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
finnhub-python==2.4.5
idna==3.3
pgcopy==1.5.0
psycopg2==2.9.1
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
